backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
gunicorn>=21.0.0
